Rename patches to get rid of Spree namespace
